#1596: LB: move elm ID to its own file
